Abstract saddleable to a separate interface
update strider interface
